Rewrite the FNFC java copy function for android
The old one was randomly throwing some crashes such as
* OutOfMemory
* NegativeArraySizeException
* IllegalStateException

This new one should fix those!

+++ source/funkin/external/android/java/funkin/extension/FNFLoaderExtension.java
@@ -2,22 +2,25 @@ package funkin.extensions;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.os.ParcelFileDescriptor;
+import android.os.Build;
+import android.os.FileUtils;
 import android.os.Bundle;
 import android.util.Log;
+import android.util.Base64;
 
 import funkin.extensions.CallbackUtil;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
 
 import org.haxe.extension.Extension;
-
-import org.haxe.lime.HaxeObject;
 
 public class FNFLoaderExtension extends Extension
 {
@@ -84,59 +87,73 @@ public class FNFLoaderExtension extends Extension
 
   public static String copyFNFCToCache(Uri uri) throws IOException
   {
-    if (uri != null)
+    if (uri == null) return null;
+
+    File output = null;
+
+    try
     {
-      String fileName = new File(uri.getPath()).getName();
-
-      if (fileName.contains(":"))
-        fileName = fileName.split(":")[1];
-
       File cacheFNFC = new File(mainContext.getCacheDir(), "fnfc");
-      File output = new File(cacheFNFC, fileName);
+      if (!cacheFNFC.exists() && !cacheFNFC.mkdirs())
+        throw new IOException("Failed to create FNFC cache dir at: " + cacheFNFC.getAbsolutePath());
 
-      if (!cacheFNFC.exists())
-        cacheFNFC.mkdir();
+      output = new File(cacheFNFC, fnfcCacheName(uri));
 
-      if (output.exists())
-        output.delete();
-
-      ParcelFileDescriptor parcelFileDescriptor = null;
-      FileInputStream fileInputStream = null;
+      InputStream in = null;
       OutputStream out = null;
 
       try
       {
-        parcelFileDescriptor = mainContext.getContentResolver().openFileDescriptor(uri, "r");
-        fileInputStream = new FileInputStream(parcelFileDescriptor.getFileDescriptor());
-
-        byte[] fileBytes = new byte[(int) parcelFileDescriptor.getStatSize()];
-        fileInputStream.read(fileBytes);
+        in = mainContext.getContentResolver().openInputStream(uri);
+        if (in == null)
+          throw new IOException("Failed to get input stream for Uri: " + uri.toString());
 
         out = new FileOutputStream(output);
-        out.write(fileBytes);
 
-        if (output.exists())
-          output.deleteOnExit();
-      }
-      catch (IOException e)
-      {
-        Log.e("trace", e.getMessage());
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q)
+        {
+          FileUtils.copy(in, out);
+        }
+        else
+        {
+          Files.copy(in, output.toPath(), StandardCopyOption.REPLACE_EXISTING);
+        }
+
+        out.flush();
       }
       finally
       {
-        if (fileInputStream != null)
-          fileInputStream.close();
-
-        if (parcelFileDescriptor != null)
-          parcelFileDescriptor.close();
-
+        if (in != null)
+          in.close();
         if (out != null)
           out.close();
       }
 
+      if (output.length() <= 0)
+        throw new IOException("The copied Uri file (" + uri.toString() + ") ended empty at: " + output.getAbsolutePath());
+
       return output.getAbsolutePath();
     }
+    catch (Throwable t)
+    {
+      Log.e("trace", "Failed to copy FNFC from " + uri.toString() + ". Error message: " + t.getMessage());
+      if (output != null) output.delete();
+      return null;
+    }
+  }
 
-    return null;
+  private static String fnfcCacheName(Uri uri)
+  {
+    try
+    {
+      byte[] digest = MessageDigest.getInstance("SHA-256")
+          .digest(uri.toString().getBytes(StandardCharsets.UTF_8));
+      return Base64.encodeToString(digest,
+          Base64.URL_SAFE | Base64.NO_WRAP | Base64.NO_PADDING) + ".fnfc";
+    }
+    catch (Exception e)
+    {
+      return Integer.toHexString(uri.toString().hashCode()) + ".fnfc";
+    }
   }
 }
